net: Move apimachinery and nncp modules to libs
These modules have no dependency on test-specific code.
Moving them to libs/net/ makes them importable from any test suite
without cross-package imports.
