+++ tests/network/localnet/ipam/conftest.py
@@ -4,8 +4,8 @@ import pytest
 from kubernetes.dynamic import DynamicClient
 from ocp_resources.namespace import Namespace
 
-import tests.network.libs.nodenetworkconfigurationpolicy as libnncp
 from libs.net import netattachdef as libnad
+from libs.net import nodenetworkconfigurationpolicy as libnncp
 from libs.net.ip import random_ipv4_address
 from libs.vm.oper import run_vms
 from libs.vm.spec import Interface, Multus, Network
